+++ tests/testthat/test-roxytest-testexamples-02-pipeline.R
@@ -45,8 +45,8 @@ test_that("[unknown alias] @ L104", {
   )
   
   # 3. Customized model -----------
-  dt = data.frame(x = rnorm(100))
-  dt$y = rpois(100, exp(1+dt$x))
+  dt <- data.frame(x = rnorm(100))
+  dt$y <- rpois(100, exp(1 + dt$x))
   m5 <- breg(dt) |>
     br_set_y("y") |>
     br_set_x("x") |>